feat: allocation pie chart based on industry  (#23)
* check

* implement get percentage allocation by industry

* edits

* delete unused file

* fix exception

* refactor pid

+++ src/main/java/com/backend/repository/PortfolioRepository.java
@@ -7,5 +7,5 @@ import com.backend.model.Portfolio;
 
 @RepositoryRestResource(exported = false)
 public interface PortfolioRepository extends JpaRepository<Portfolio, Long> {
-    Portfolio findByPid(int pid);
+    Portfolio findByPid(long pid);
 }
